Rename `processCircuits` file `circuits`
The name `processCircuits` does not fit the convention for other modules.
